task_3: исправил ошибки detekt
