+++ app/src/main/java/ru/practicum/android/diploma/ui/about/AboutFragment.kt
@@ -18,7 +18,5 @@ class AboutFragment : BindingFragment<FragmentAboutBinding>() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
