+++ app/src/main/java/ru/practicum/android/diploma/ui/favorite/FavoriteFragment.kt
@@ -18,7 +18,5 @@ class FavoriteFragment : BindingFragment<FragmentFavoriteBinding>() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
