+++ app/src/main/java/ru/practicum/android/diploma/ui/filter/FilterSettingsFragment.kt
@@ -1,24 +1,22 @@
-package ru.practicum.android.diploma.ui.region_selection
+package ru.practicum.android.diploma.ui.filter
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.databinding.FragmentRegionSelectionBinding
+import ru.practicum.android.diploma.databinding.FragmentFilterSettingsBinding
 import ru.practicum.android.diploma.ui.fragment.BindingFragment
 
-class RegionSelectionFragment : BindingFragment<FragmentRegionSelectionBinding>() {
+class FilterSettingsFragment : BindingFragment<FragmentFilterSettingsBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentRegionSelectionBinding {
-        return FragmentRegionSelectionBinding.inflate(inflater, container, false)
+    ): FragmentFilterSettingsBinding {
+        return FragmentFilterSettingsBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
