+++ app/src/main/java/ru/practicum/android/diploma/ui/search/SearchFragment.kt
@@ -18,7 +18,5 @@ class SearchFragment : BindingFragment<FragmentSearchBinding>() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
