+++ app/src/main/java/ru/practicum/android/diploma/ui/selections/country/CountrySelectionFragment.kt
@@ -1,24 +1,22 @@
-package ru.practicum.android.diploma.ui.similar_vacancy
+package ru.practicum.android.diploma.ui.selections.country
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.databinding.FragmentSimilarVacancyBinding
+import ru.practicum.android.diploma.databinding.FragmentCountrySelectionBinding
 import ru.practicum.android.diploma.ui.fragment.BindingFragment
 
-class SimilarVacancyFragment : BindingFragment<FragmentSimilarVacancyBinding>() {
+class CountrySelectionFragment : BindingFragment<FragmentCountrySelectionBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentSimilarVacancyBinding {
-        return FragmentSimilarVacancyBinding.inflate(inflater, container, false)
+    ): FragmentCountrySelectionBinding {
+        return FragmentCountrySelectionBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
