+++ app/src/main/java/ru/practicum/android/diploma/ui/selections/industry/IndustrySelectionFragment.kt
@@ -1,24 +1,22 @@
-package ru.practicum.android.diploma.ui.filter_settings
+package ru.practicum.android.diploma.ui.selections.industry
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.databinding.FragmentFilterSettingsBinding
+import ru.practicum.android.diploma.databinding.FragmentIndustrySelectionBinding
 import ru.practicum.android.diploma.ui.fragment.BindingFragment
 
-class FilterSettingsFragment : BindingFragment<FragmentFilterSettingsBinding>() {
+class IndustrySelectionFragment : BindingFragment<FragmentIndustrySelectionBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentFilterSettingsBinding {
-        return FragmentFilterSettingsBinding.inflate(inflater, container, false)
+    ): FragmentIndustrySelectionBinding {
+        return FragmentIndustrySelectionBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
