+++ app/src/main/java/ru/practicum/android/diploma/ui/selections/region/RegionSelectionFragment.kt
@@ -1,24 +1,22 @@
-package ru.practicum.android.diploma.ui.industry_selection
+package ru.practicum.android.diploma.ui.selections.region
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.databinding.FragmentIndustrySelectionBinding
+import ru.practicum.android.diploma.databinding.FragmentRegionSelectionBinding
 import ru.practicum.android.diploma.ui.fragment.BindingFragment
 
-class IndustrySelectionFragment : BindingFragment<FragmentIndustrySelectionBinding>() {
+class RegionSelectionFragment : BindingFragment<FragmentRegionSelectionBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentIndustrySelectionBinding {
-        return FragmentIndustrySelectionBinding.inflate(inflater, container, false)
+    ): FragmentRegionSelectionBinding {
+        return FragmentRegionSelectionBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
