+++ app/src/main/java/ru/practicum/android/diploma/ui/similar/SimilarVacancyFragment.kt
@@ -1,24 +1,22 @@
-package ru.practicum.android.diploma.ui.country_selection
+package ru.practicum.android.diploma.ui.similar
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.databinding.FragmentCountrySelectionBinding
+import ru.practicum.android.diploma.databinding.FragmentSimilarVacancyBinding
 import ru.practicum.android.diploma.ui.fragment.BindingFragment
 
-class CountrySelectionFragment : BindingFragment<FragmentCountrySelectionBinding>() {
+class SimilarVacancyFragment : BindingFragment<FragmentSimilarVacancyBinding>() {
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentCountrySelectionBinding {
-        return FragmentCountrySelectionBinding.inflate(inflater, container, false)
+    ): FragmentSimilarVacancyBinding {
+        return FragmentSimilarVacancyBinding.inflate(inflater, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
