+++ app/src/main/java/ru/practicum/android/diploma/ui/vacancy/VacancyFragment.kt
@@ -18,7 +18,5 @@ class VacancyFragment : BindingFragment<FragmentVacancyBinding>() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-
     }
 }
